songR: Self-Organizing Nebulous Growths for Dimensionality Reduction

Native R/C++ implementation of the Self-Organizing Nebulous Growths (SONG) algorithm for nonlinear dimensionality reduction and data visualization. SONG is a parametric method that supports incremental addition of new data to existing visualizations without reinitialization, handles both homogeneous and heterogeneous data increments, and is robust to noise and highly mixed clusters. Based on the algorithm described in Senanayake et al. (2021) doi:10.1109/TNNLS.2020.3023941 . Includes an interactive Shiny application for comparing SONG, t-SNE, and UMAP visualizations side-by-side.

References

Senanayake, D. A., Wang, W., Naik, S. H., & Halgamuge, S. (2021). Self-Organizing Nebulous Growths for Robust and Incremental Data Visualization. IEEE Transactions on Neural Networks and Learning Systems, 32(10), 4588–4602. doi:10.1109/TNNLS.2020.3023941

See also

Author

Maintainer: Raban Heller raban.heller@uni-ulm.de (ORCID)